Initial support for sharing holes between fields on the same plate. Current use case is multiple fields with common set of skys or G&A
Improvements to assignment algorithm. fixed bugs in LR rejiggering and simplified process. Support for holes that are shared between setups (e.g. two fields share skys and will be assigned together)

Fairly major overhaul of minsky now takes T/F (same behavior as before) or a value which is the priority above which things must be kept. This value does not affect skys (assign._assign_fibers:255). This effectively obsoletes keepall, which no longer affects anything, but hasn't been removed from the code.

Added __eq__ to hole so it is properly hashable, that fixed a bug that appeared in the process